backend validations

+++ api/src/app/controllers/ContactController.js
@@ -1,6 +1,6 @@
 /* eslint-disable */
 const ContactRepository = require('../repositories/ContactsRepository');
-const isValidUUID = require('../utils/isValidUUID')
+const isValidUUID = require('../utils/isValidUUID');
 class ContactController {
 	async index(request, response) {
 		// Listar todos os registros
@@ -16,7 +16,7 @@ class ContactController {
 		const { id } = request.params;
 
 		if (!isValidUUID(id)) {
-			return response.status(400).json({ error: 'Invalid user id' })
+			return response.status(400).json({ error: 'Invalid user id' });
 		}
 		const contact = await ContactRepository.findById(id);
 
@@ -36,18 +36,24 @@ class ContactController {
 			return response.status(400).json({ error: 'Name is required' });
 		}
 
-		const contactExists = await ContactRepository.findByEmail(email);
-		if (contactExists) {
-			return response
-				.status(400)
-				.json({ error: 'This email is already in use' });
+		if (category_id && !isValidUUID(category_id)) {
+			return response.status(400).json({ error: 'Invalid category' });
+		}
+
+		if (email) {
+			const contactExists = await ContactRepository.findByEmail(email);
+			if (contactExists) {
+				return response
+					.status(400)
+					.json({ error: 'This email is already in use' });
+			}
 		}
 
 		const contact = await ContactRepository.create({
 			name,
-			email,
+			email: email || null,
 			phone,
-			category_id,
+			category_id: category_id || null,
 		});
 		response.status(201).json(contact);
 	}
@@ -58,27 +64,34 @@ class ContactController {
 		const { name, email, phone, category_id } = request.body;
 
 		if (!isValidUUID(id)) {
-			return response.status(400).json({ error: 'Invalid user id' })
+			return response.status(400).json({ error: 'Invalid content id' });
+		}
+		if (!name) {
+			return response.status(400).json({ error: 'Name is required' });
+		}
+
+		if (category_id && !isValidUUID(category_id)) {
+			return response.status(400).json({ error: 'Invalid category' });
 		}
 
 		const contactExists = await ContactRepository.findById(id);
 		if (!contactExists) {
 			return response.status(404).json({ error: 'Contact not found' });
 		}
-		if (!name) {
-			return response.status(400).json({ error: 'Name is required' });
-		}
-		const contactByEmail = await ContactRepository.findByEmail(email);
-		if (contactByEmail && contactByEmail.id !== id) {
-			return response
-				.status(400)
-				.json({ error: 'This email is alkready in use' });
+
+		if (email) {
+			const contactByEmail = await ContactRepository.findByEmail(email);
+			if (contactByEmail && contactByEmail.id !== id) {
+				return response
+					.status(400)
+					.json({ error: 'This email is alkready in use' });
+			}
 		}
 		const contact = await ContactRepository.update(id, {
 			name,
-			email,
+			email: email || null,
 			phone,
-			category_id,
+			category_id: category_id || null,
 		});
 		response.json(contact);
 	}
@@ -86,6 +99,9 @@ class ContactController {
 	async delete(request, response) {
 		// Deletar um registro
 		const { id } = request.params;
+		if (!isValidUUID(id)) {
+			return response.status(400).json({ error: 'Invalid content id' });
+		}
 
 		await ContactRepository.delete(id);
 		// 204 :No content
